+++ dev/ext/lib/TimeHandling/EpochDataStore.cpp
@@ -225,7 +225,7 @@ namespace gpstk
          ite--;
       }
       
-      const int N = its->second.size();
+      const size_t N = its->second.size();
 
       std::vector<double> times;
       std::vector<std::vector<double> > datas(N);
@@ -247,7 +247,7 @@ namespace gpstk
             
          }
          
-         for(int i = 0; i < N; i++)
+         for(size_t i = 0; i < N; i++)
          {
             datas[i].push_back( vd[i] );
          }
@@ -257,7 +257,7 @@ namespace gpstk
 
       double dt = t - its->first;
       
-      for(int i = 0; i < N; i++)
+      for(size_t i = 0; i < N; i++)
       {
          vd[i] = SimpleLagrangeInterpolation(times,datas[i],dt);
       }
